requirements.txt
torch>=2.3.1
torchvision>=0.18.1
Pillow>=10.3.0
numpy>=1.26.4